Changed the value types of user input, expected answer and generation numbers variables to prevent crashes Also added new comment line

+++ SimpleMathGameV2/MGSimpleHead.cpp
@@ -83,7 +83,7 @@ void Problem::Display()
 	printf_s("%d %c %d = ", nFirstI, cCheckActionSymbol(), nSecondI);
 }
 
-int Problem::nExpectedAswer()
+long long Problem::nExpectedAswer()
 {
 	switch (lAction)
 	{
